requirements.txt
Flask==2.3.3
groq>=0.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7